Fix implementation issue with `fseek reg` Change hyphen to underscore in `tape-device` as `tape-device` isn't a valid rust identifier Fix some formatting Correct and update language
