+++ src/constants.rs
@@ -276,18 +276,15 @@ pub const ALL_OPS: [u8; 109] = [
     PRT_REG, PRT_VAL, PRT_AREG,
     PRTLN,
     PRTC_REG, PRTC_VAL, PRTC_AREG,
-    FOPEN_REG,
+    FOPEN_VAL, FOPEN_REG,
     FILER_REG_ADDR, FILER_REG_AREG,
     FILEW_REG_ADDR, FILEW_REG_AREG,
-    FSEEK_REG,
-    FSKIP_REG_REG,
+    FSEEK_REG, FSEEK_VAL,
+    FSKIP_VAL_REG, FSKIP_REG_REG,
     FCHK_REG_ADDR, FCHK_REG_AREG,
-    FOPEN_VAL,
+    FCHK_VAL_ADDR, FCHK_VAL_AREG,
     FILER_VAL_ADDR, FILER_VAL_AREG,
     FILEW_VAL_ADDR, FILEW_VAL_AREG,
-    FSEEK_VAL,
-    FSKIP_VAL_REG,
-    FCHK_VAL_ADDR, FCHK_VAL_AREG,
     NOP,
     HALT,
     ARG_REG_VAL, ARG_REG_REG,
